Changes to Truth's DoubleSubject's handling of non-finite values.
The main behavioural changes is that an isNotWithin assertion now fails if either the subject or the object is either infinite. (Previously, it would pass if one argument was infinite and the other was finite, or if one argument was POSITIVE_INFINITY and the other argument was NEGATIVE_INFINITY. It already fails if either argument is NaN. An isWithin assertion already fails if either argument is either infinite or NaN.)

This makes the corresponding change to MathUtil.notEquals(double, double, double). It also fixes MathUtil.notEquals(float, float, float) which had completely inverted the logic (it was untested, and luckily is also unused).

Additionally, some of the error messages are improved.

Also specify the existing and new behaviours in the javadoc, and add tests. And add tests for the non-special failing cases, since they seemed to be missing.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=98196969