Re-work the failure message storage so that it is not stored in the Verb
Previously it was stored in AbstractVerb, and accessed by a delegating failure message which stored the TestVerb.  There's no real reason to keep a TestVerb around when we already have a new instance of FailureStrategy which delegates, so this change moves the storage of the overriding failure message to the delegating FailureStrategy instance.

This fixes issue #199 where check() doesn't preserve the failure message, because it doesn't have proper access to the original verb.

While in the area added proper tests for StringUtil, and make it properly throw when there are too many placeholders.