Fix four unchecked cast warnings in MapSubject and MultimapSubject. They seem easy enough to avoid, so we might as well.
Two of them are trivial: Entry<?, ?> works just as well as Entry<Object, Object> and there is no need for the cast.

In the other two, I'm converting multimap.get(key) to multimap.asMap().get(key) inside an if (multimap.containsKey(key)) condition, which is guaranteed to return the same thing, and again avoids the cast. The alternative would be to suppress the warning, since I think it's sort of implicitly safe by specification, but since we probably don't care about the possible performance cost of the asMap(), this seems neater. Happy to switch if you'd prefer.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=129111862