Renames `getSubject()` -> `actual()` and `getDisplaySubject()` -> `actualAsString()`
Also adds `actualCustomStringRepresentation()` to let subjects themselves set a "displayable actual" rather than the unhappy practice of overriding the (formerly named) `getDisplaySubject()` which not only has a string representation of the content of the actual value, but has logic related to `.named()` and other contextual formatting.

Preserves the old methods but deprecates them, to be cleaned up after.

This would normally change semantics in a couple of cases where subjects were overriding getDisplaySubject() deliberately to alter the semantics of .named() to be a replacing operation rather than a prefixing one.  A temporary backwards compatibility is hacked in, but should be removed once a truly replacing "displayedAs()" method is added in a later commit.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=132085665