Fix a typo (I assume) in one of the failure messages from IterableSubject.containsExactly: it says "contains only" rather than "contains exactly".
I've done a bit of random tidy-up while I was here. For one thing, the private method took a failVerb parameter which always had the same value, so I eliminated that. And that method is more similar to containExactlyElementsIn than it is to containsExactly, so I renamed it (which also means there's no confusion between the private two-arg overload the the public varargs overload).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=135104047