Add IterableSubject.ComparingElementsUsing.containsExactly. It just delegates to containsExactlyElementsIn.
Note that it doesn't need to handle the "caller accidentally did containsExactly with a single Iterable argument when they meant containsExactlyElementsIn" test that the regular containsExactly does, because the generic type signature won't allow it. It does handle the null array case the same way.

This CL includes a reduced set of tests, relative to the extensive set for containsExactlyElementsIn, which seems okay since it is a simple delegation. It does add tests with nulls for the existing contains and containsExactlyElementsIn methods.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=135481756