Remove MultimapSubject.UsingCorrespondence.valuesForKey.
We did agree to do this in the API review, but on reflection I'm not sure whether it's a good idea. We already support

    assertThat(actual)
        .valuesForKey("me")
        .comparingElementsUsing(STRING_PARSES_TO_INTEGER_CORRESPONDENCE)
        .contains(456);

and implementing this method would allow you to write

    assertThat(actual)
        .comparingValuesUsing(STRING_PARSES_TO_INTEGER_CORRESPONDENCE)
        .valuesForKey("me")
        .contains(456);

which is equivalent. We think that having two ways of spelling the same thing is  unhelpful, because it risks creating confusion.

Additionally, the former snippet (the one that already exists) allows you to call named() on the result of valuesForKey; the latter (which we'd get if we implemented this method) does not. It seems especially unhelpful if we say "there are two valid ways of doing this *unless* you want that named, in which case you have to pick the right one, which isn't exactly obvious".

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=144303806