Bring the truth deps up to date
Also, remove the "truth.version" property and explicit version setting in the extensions.  They should be relying on ${project.version} so as to ensure that when the project's version is updated, it inherits this value as the version of truth upon which it also depends.  This should avoid any future occasions where we publish a release version with a dependency on a truth SNAPSHOT version.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=148281908