Set doc and window titles.
Currently they say "Truth (Parent) ..." because that's (sensibly enough, I would assume) the name of the parent pom.
(The reason that Maven takes the name of the parent pom, rather than the core pom, is that we generate Javadoc by running javadoc:aggregate on the parent (so that we include the main extensions in the same Javadoc "site" as the core).)

Also, add a project name to the core pom.xml. I'd initially hoped that that would be enough to set the Javadoc title, but it wasn't for the reason described above. Still, it seems nicer to say "Truth Core" than "truth" in the build output, especially since we have proper names in the other poms.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=160994277