Annotate empty strings in collections printed by IterableSubject.
Example error message:

  Not true that <[]> contains exactly <["" (empty String)]>. It is missing <["" (empty String)]>

=== Note: ===

Some methods such as `contains()` and `containsAnyOf()` can't be changed. The reason is that their error message has the form "Not true that <[]> contains any of <[]>" where the first list is actualAsString(). Changing only the second list could lead to confusion because sometimes it'd show "" (empty String) and sometimes it wouldn't.

Consequently, I adopted the criterion of only changing error messages where I can at annotate at least the expected list and some derived list, such as "Not true that <[]> contains all of <[]>. It is missing <[]>".

Obviously, this would be solved if we stop using actualAsString(), but IIRC this is not currently the plan.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=163937787