Automated g4 rollback of changelist 165629291.
*** Reason for rollback ***

Breaks expected-exception tests that expect the original exception to be propagated.

*** Original change description ***

Change Expect to catch an exception thrown during the evaluation of a Statement and treat is as a failure.

Change ExpectationGatherer::toString to use stripTruthStackTrace on any failure causes.

Background:
Prior to this change, if an exception was thrown in a test, any previous failures gathered up to that point are not printed, only the exception that was thrown during Statement evaluation was shown in test output. This can lead to misleading and/or confusing test output. Additionally, stack...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=165639209