Link from Expect to ExpectFailure.
We already have a link in the ExpectationGatherer deprecation text, but the link will go away when the class is made public. And it's clear from the old ExpectationGatherer user and from CL 167017292 that people are finding Expect when they want ExpectFailure.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=167026073