Remove some more low-value test code.
The `delegation` files are a less fleshed-out version of the Employee example.
ExampleTest is more fleshed out but depends on `delegation`, and I'd rather remove it than modify it.
BarSubject might once have been part of a test for repeated subjects, but we ripped that out a while back.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=167996710