Set -source 1.6 -target 1.6 on *all* modules, not just the core.
javac9 is removing support for -source 1.5 -target 1.5, which is what the Maven compiler plugin we're using currently defaults to. (I'm kind of surprised that the build was working at head, given that the other modules are compiled against the core, which already was built for 1.6.)
(Presumably newer version of the compiler plugin will have a newer default, but let's get out ahead of things. And given that we already pick a version for the core, we might as well lock it in everywhere.)

Also, move the existing compiler configuration from <plugins> to <pluginManagement>. The docs seem to suggest that that's the more sensible thing, and it's what we do with our other plugins. (I'm actually slightly confused about the difference between the two as described in the docs, probably because I'm too lazy to think about it in detail right now.)

Finally, move the EmployeeSubjectTest exclusion into the core, since that's the only module it applies to.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=170081872