A small and pedantic refactoring of IterableSubject.
This just makes sure that the containsExactly methods return ALREADY_FAILED instead of IN_ORDER if the expected iterable is empty and the actual iterable is not. This makes no practical difference, since both implementations are identical, but since the whole point of having two differently named noops is to make the code self-documenting, we might as well take the care to get it right.

In the fuzzy case, the special-casing moves from correspondInOrderExactly to its caller. This improves readability, since previously correspondInOrderExactly would have the undocumented side-effect of failing and then (if that didn't throw) misleadingly return true in the "should be empty but isn't" case.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=178387356