Improve the failure messages from MapSubject.UsingCorrespondence.containsExactly and containsExactlyElementsIn.
This is the Fuzzy Truth version of [] This change also fixes an unintended regression from that CL in the failure message from a non-fuzzy containsExactly with empty expected map, and adds tests for that case.

It also fixes the only place the TGP showed was asserting on the fuzzy failure message.

Unfortunately, we have to switch from using guava's MapDifference to rolling our own, because guava's requires the Maps to have the same value types and to be tested with a (symmetric) Equivalence, so we can't use it without relying on undocumented implementation details. (It would be very little work to upgrade the guava one, but it would be a breaking change.)

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=178644906