Remove another low-value test.
If check() were to start dropping FailureStrategy, a large number of our existing tests would fail, as they'd stop using ExpectFailure's capturing logic and start throwing AssertionErrors.
Additionally, I'm not sure that this test even does check that the FailureStrategy persists: It appears to be checking that check().that(...).... throws an AssertionError, which might mean it's preserving the default or might mean someone hardcoded it to the wrong thing.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=179956444