Store the full chain of Subject instances.
This doesn't accomplish anything on its own. Rather, it makes things easier for future CLs. It turns out to be a lot simpler to track everything and decide what we need later, rather than try to make that decision on the fly for each of n different fields we might have in this class someday. If we do see performance problems, we can be smarter about it.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=180193272