Start implementing Smart Diffs for Fuzzy Truth.
This CL adds the formatDiff method to Correspondence, and uses it in the specific case of iterables with containsExactly with one missing and one unexpected.

The displayingDiffsPairedBy method and the use of formatDiff in other assertions will follow in later CLs.

It seems worth landing this much, since this the case covered here is exactly the one that several people were asking to be handled on the bug. (They won't get any advantage until Proto Truth starts implementing formatDiff, but we're still moving in the right direction.)

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=182394375