Start work on displayingDiffsPairedBy for Smart Diffs for Fuzzy Truth.
This adds the methods and stores the specified values but doesn't do anything with them. Since the methods are useless, they're package visibility for now. (I figured it would be useful to review the APIs and the documentation separately from reviewing the implementation and the choice of failure message.)

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=184553216