Introduce a Field class and a method to format fields into a failure message.
This lays some groundwork for the coming changes to failure messages, in which Subject and FailureMetadata will pass a list of fields to construct their failure exceptions.

Also, start using Field in a trivial way in FailureMetadata: Instead of returning @Nullable String, return Optional<Field>, and convert it to a String as needed. This sets up for using the Field directly in future commit. (Eventually, the old-style failure messages will go away, and we can eliminate the String methods entirely.)

expected: abc
but was : xyz

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=186516409