Make IterableSubject.UsingCorrespondence.displayingDiffsPairedBy() work for containsExactly and containsExactlyElementsIn, and make it public.
This follows the precedent set in ef55cb75d588f8af387f69b39e6aac0259215b42 for containsNoneOf/In and says "corresponds to" instead of repeating the correspondence.toString() when in a loop. (I actually vaguely think that we should switch to always saying "corresponds to" on the second and subsequent occurrences. That would be possible, but it would create churn in several existing messages, so I'm not doing it here. At least, though, we should make sure we never repeat correspondence.toString() a number of times that scales with the size of the input.)

Also adds a TODO in fuzzy.md to remind me to update it in due course. (My plan is to do that when the feature is more-or-less complete, and to put something in the RELNOTES for that commit which can direct folks to the docs for more info.)

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=186598186