Put explicit null handling in our equality checking.
By getting the null checks out of the way early, we make it easier to insert tests for array equality later (https://github.com/google/truth/issues/335).

As part of this, we need to continue to explicitly check actual()==expected to avoid breaking various tests that fail if we actually invoke equals().
Furthermore: The usual way to do this would be the check actual()==expected at the top of the method. However, this seems to cause j2cl to consider 0 to be == to 0.0. For whatever reason, I'm able to avoid it by moving the == check down.

(Also, make a tiny simplification by changing integralValue() to return a primitive long rather than a Long. That lets us use == instead of equals().)

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=187177643