More expectFailureWhenTesting() helpers.
Besides simplifying things for null, the helper also prevents us from needing 3 lines (expectFailure, whenTesting, that) when google-java-format puts each call on its own line.

It also identified a least one test (and maybe others; I forget) that wassn't testing what it appears to be: ComparableSubjectTest.isEquivalentAccordingToCompareTo() was testing BigDecimalSubject, not ComparableSubjectTest.

Finally, this prompted me to remove a bunch of duplicate tests in IntegerSubjectTest, to reorganize it to match LongSubjectTest, and to move a few tests from the former to the latter.

(I grant that, for some tests, there's little to no improvement, but it didn't seem worth trying to pick and choose where to add the helper and where not to, especially since it's hard to predict when we'll have bugs like the one in ComparableSubjectTest.)

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=188040561