Make IterableSubject.UsingCorrespondence.displayingDiffsPairedBy() work for containsAnyOf/In().
This is the final change for this feature, apart from documentation... and multimaps, where the behaviour is TBD.

(The text of the failure messages is hopefully 'good enough' for now. I don't think it's worth fiddling with the punctuation given the upcoming changes. It's quite wordy, but it does happen only if the caller specifically asks for it.)

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=189761499