Make factValue() reject facts that have keys but no value.
It feels icky to me that anyone would pass null as a magic value, and there are already ways to check that a given key is present.
Technically there's no longer a way to test that a given key has no associated value, but it is hard to imagine that people would get this wrong, write tests for it, and fix the bug as a result. (It hasn't come up in Truth itself, at least.) And we can add support later if needed -- always easier than removing support after the fact.

Also, add tests for facts that have keys without values.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=190790320