Change all containsExactly() fail() calls to go through a helper method.
This reduces duplication a bit, and it should be more natural when we switch to the key-value message format.

Also, IntelliJ static imported some methods everywhere when I chose to static import in one place, so that's in this CL, too, because why not?

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=191443280