Migrate off failComparing(), which is being deprecated.
Also, remove a branch that could never be taken. (We've already verified above that the classes match.)

And reduce nesting just slightly, too.

As for failComparing(): In most cases that I saw, the change was roughly:

Before:
  if (actual().foo() != expectedFoo) {
    failComparing("foo", actual().foo, expectedFoo);
  }

After:
  check("foo()").that(actual().foo()).isEqualTo(expectedFoo);

When it works out, it's shorter, and it produces messages that fit with Truth's other equality assertions. (It also makes it easier to get the order of the actual and expected parameters right. I've fixed such bugs in another CL.)

Here, we still need the equality check, so that code doesn't end up any shorter from migrating off failComparing(). What we do get out of the migration (for better or for worse :)) is that it moves us off the old failure API and message format.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=192275122