Reimplement most of the old failure API in terms of the new, which sometimes adds a "name:" field.
This ensures that our forthcoming migration CLs don't change the failure message (though again, this CL itself can change it slightly).
(I haven't reimplemented failWithRawMessageAndCause or failComparing. I've already migrated their users.)

The change to failure messages: If an assertion uses named(), this CL adds a "name:" field to the beginning. This is sometimes good and sometimes bad:
- The good part is that some subjects currently ignore calls to named(). That is, if someone calls named("foo"), "foo" doesn't appear in the failure message. In particular, failWithRawMessage doesn't currently add the name to the failure message, so subject authors are responsible for adding it themselves. The large majority do not. (Have I mentioned that I want to delete named()? :))
- The bad part is that some subjects already include named() in their message -- whether by manually adding it when calling failWithRawMessage or by calling one of the other old fail* methods, which include the name automatically. Users of these subjects will the name twice -- once in the new "name:" field and once in the message body. This is unfortunate, but it should be temporary: First, as noted above, I hope to delete named(). And second, even if I don't, we're still looking to migrate users to the new failure API, and that API adds the name only once, rather than twice as some of the old API will do as of this CL. Sorry for the ugly messages in the meantime.

(The new API, I should note, is still not publicly available, as we have a few final decisions to nail down.)

Also, fix a minor bug in Fact: We were adding trailing whitespace to the end of "simple" keys so that their values would align. This makes no sense because they have no values. The whitespace is mostly harmless but is confusing if you're writing tests that check for an exact failure message.

RELNOTES=Changed the old API's failure messages to automatically add a "name: ..." field if the user called `named()`. In some cases, this adds a name where it was missing before; in others, it duplicates a name that is already present. The long-term fix for this will be to migrate to the new failure API, which always includes the name once, and possibly also to delete `named()` in favor of `withMessage()`.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=195091598