Refactor ShouldIgnore's {YES,NO,MAYBE} into {YES,NO}x{RECURSIVE,NONRECURSIVE}.
This is, I think, a lot simpler to comprehend than 'MAYBE' was before.  It simplifies the FieldScopeLogic classes and gives ProtoTruthMessageDifferencer a clearer path to deciding 'should this field and/or subtree be ignored?'.  Both classes (ProtoTruthMessageDifferencer and FieldScopeLogic) are now speaking the same language with regards to recursiveness.

The reason this fixes the bug is because 'MAYBE' was being used to mean mostly 'YES_NONRECURSIVE', but the NegationFieldScope was evaluating !MAYBE as MAYBE, which is mostly okay but leads to incorrect behavior in this specific case (empty submessage in one proto, not set in another).  Now that 'YES_NONRECURSIVE' has a proper negation, the incorrect behavior is gone.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=199193103