Remove the fail* APIs that we've deprecated so far. Also, reimplement StandardSubjectBuilder.fail() to use Fact instances.
Then clean up a bunch of stuff:
- We no longer need to support JUnit ComparisonFailure messages, including hacking a suffix onto them.
- We no longer need to provide the description and root object as an Optional/String in addition to as an ImmutableList<Fact>.
- I nailed down what parameters could be null (just `cause`, IIRC).

(Also, the change to StandardSubjectBuilder.fail() lets it align fact values if people call check().fail(), unlikely as that is.)

Plus:

Migrate off failWithRawMessageAndCause, which is being deprecated.

I'm sorry that the replacement is so ugly. Our reason for not supporting this directly is partially that it's needed by only ~1% of subjects. But mostly, we've over time taken the position that Truth is intended for only the "assert" phases of the "arrange, act, assert" testing process[*]. Exceptions, on the other hand, usually come from the earlier phases.

Unfortunately, I still owe everyone a long explanation of *why* we have come to this position. If it's bugging you now, let me know, and I can make it happen.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=200405939