Refactor ShouldIgnore (again) into FieldScopeResult, using names more generally about inclusion and exclusion in a scope rather than ignores.
This is to facilitate follow-up CLs where we'll be using more FieldScopeLogics for other things, not just for whether particular values should be compared.  FluentEqualityConfig.fieldScopeLogic() will not be the only FieldScopeLogic on the config either, so it is renamed here.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=203029565