Get dependencyManagement working for Truth's own modules.
The problem before was the "circular" Truth[tests] -> Compile-Testing -> Truth dependency chain. Because we started using dependencyManagement, Maven wanted Compile-Testing to depend on the specific version of Truth that we specified there. That meant that Truth 1.0-SNAPSHOT needed to be available before we could build Truth 1.0-SNAPSHOT. (Again, not really, because Compile-Testing needs only Truth itself, and Truth doesn't need Compile-Testing until its _tests_, but that's circular to Maven. We had similar problems with guava and guava-testlib, IIRC, and that's why we have a separate guava-tests artifact.)

Anyway, now that we exclude Compile-Testing's dependency on Truth, Maven doesn't go looking for a version of Truth that we haven't built yet.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=210937639