Suppress warnings in classes that implement equals() without also implementing hashCode().
The contract for Object.hashCode states that if two objects are equal, then
calling the hashCode() method on each of the two objects must produce the same
result. Implementing equals() but not hashCode() causes broken behaviour when
trying to store the object in a collection.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=218007454