Suppress an ErrorProne warning in Truth.
The UndefinedEquals warning[*] triggers for MultimapSubject.ieEqualTo, because it calls Objects.equal on the multimaps. Whatever the issues with the contract of Multimap.equals, MultimapSubject.isEqualTo is required to follow that contract, so suppressing the warning seems like the right thing to do.

[*] https://github.com/google/error-prone/blob/master/docs/bugpattern/UndefinedEquals.md

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=227852973