Add a separator to the end of the stack trace when reporting exceptions caught from Correspondence.compare etc., and add tests.
The separator so that when this stack trace appears at the end of an AssertionError message (which it generally does), the stack trace of the stored exception is distinguishable from the stack trace of the AssertionError. Without this, it's pretty confusing.

The added tests would have caught this issue fixed in c4b5ab73c260cbd0a27cfa63b3aeec7631ec5291 (although they're still a little permissive, the chances that the rest of the stack trace are wrong seem pretty slim).

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=228378358