Make all Fuzzy Truth assertions in IterableSubject handle exceptions from Correspondence.formatDiff.
A follow-up CL will do the same for MapSubject. (MultimapSubject doesn't use formatDiff yet.)

This CL refactors how Correspondence.ExceptionStore works a bit, to make its call-sites neater. Previously, you had one ExceptionStore instance for compare() exceptions, and we'd have had to add a second for formatDiff() exceptions, and later a third for pairing exceptions, and we'd have had to pass up to three instances around everywhere. (Additionally, it was a bit weird that you said that the instance was forCompare() when you created it, but then you had to repeat the name of the method when you added an exception to it.) With this refactoring, there is one ExceptionStore instance, and you say which method is throwing the exception when you add it, and it files them appropriately. So we only have to pass one of these things around the various layers.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=229530243