Copy some GWT settings from Guava to Truth.
The Truth GWT Travis build is failing with "GC overhead limit exceeded":
https://travis-ci.org/google/truth/builds/498829960

I don't know whether this CL will help or not. If not, the next thing to try is probably to define a module and set it with <module>, as we also do with Guava.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=235764378