Include the reference to Proto Truth in the github version of fuzzy.md.
I can't see any reason for this to be Google-internal any more.

Also change the link at the top of the page to refer to the Proto Truth docs (which refer to the protobuf docs) instead of to the protobuf docs (since this seems more helpful).

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=238503042