Deprecate the no-arg Subject.check().
We've migrated users to the other overload when it was practical to automate, and we've manually cleaned up the remaining third-party users to convince ourselves that there's no especially low-hanging fruit left. Users don't need to migrate off this overload, but we can try to prevent it from attracting more usages. (We also have a code-review-time check that suggests a replacement in the simple case.)

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=247061298