Make ProtoTruth Subjects extend IterableSubject, etc.
This saves us from redeclaring a bunch of methods, and it's something we've considered doing for a while. But the immediate motivation is to eliminate our use of no-arg check(). There are other ways we could do that, but this larger project felt worth addressing now.

I've also made some other changes to avoid the no-arg check() by storing a FailureMetadata instance (usually a bad idea, but OK here for reasons documented in the code). This effectively undoes parts of CL 170792474.

[]

[]

Relnotes:
  - Made `MapSubject` and `MultimapSubject` extensible.
  - Made ProtoTruth `Subject` classes extend `IterableSubject`, `MapSubject`, and `MultimapSubject`.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=250565170