Update to JDK11.
This is an attempt to fix the error:
  Expected feature release number in range of 9 to 13, but got: 8
  The command "~/bin/install-jdk.sh --target "/home/travis/oraclejdk8" --workspace "/home/travis/.cache/install-jdk" --feature "8" --license "BCL"" failed and exited with 3 during .
https://travis-ci.org/google/truth/builds/544847563

Possibly we could just request an old version of Ubuntu:
  dist: trusty
https://travis-ci.org/google/truth/builds/544847563

But we'll have to update eventually, so let's try now.

If this works, I can replicate it to our other projects.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=253584474