Remove workaround for broken Map.isEqualTo() implementations.
We've decided that Truth isn't in the business of handling the "act" part of "arrange, act, assert." To test an implementation of equals(), users can call equals() themselves.

But that was the case way back when we originally added this. What really prompted this now is that I realized this change would chip away at "Not true that..."-style messages :) Additionally, I'm looking at making a similar change to IterableSubject, and I figure this is as good a time as any to be consistent in not handling bad equals() methods (which we already don't handle in MultimapSubject).

Also, avoid NPE on null actual value.

Relevant: https://github.com/google/truth/issues/268

[]

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=253994206